refactor: change mod name
