+++ runtime/src/block_job/mod.rs
@@ -1,10 +1,10 @@
-mod block;
 mod input;
+mod job_handle;
 mod listener;
 mod service_job;
 mod task_job;
 
-pub use block::BlockJobHandle;
 pub use input::{fulfill_nullable_and_default, validate_inputs};
+pub use job_handle::BlockJobHandle;
 pub use service_job::{execute_service_job, ServiceJobParameters};
 pub use task_job::{execute_task_job, TaskJobParameters};
